Phase 7.1c: Migrate keyboard shortcuts components to plain CSS
- Migrated ShowKeyboardShortcuts.tsx with 8 styled sub-components
- Migrated KeyboardShortcutsPopup.tsx with StyledModal wrapper
- Created plain CSS versions with CSS variable theming
- Created .legacy.tsx files for backward compatibility
- Maintained original export paths via updated .tsx files
- Converted all theme.breakpoints.mobile() to plain CSS media queries
- Added proper theme prop filtering to prevent DOM leakage
