+++ src/app/content/keyboardShortcuts/components/KeyboardShortcutsPopup.css
@@ -1,0 +1,7 @@
+/**
+ * KeyboardShortcutsPopup - Plain CSS version
+ */
+
+.keyboard-shortcuts-popup-modal {
+  max-width: 92.8rem;
+}
